# Conflicts:
#	catm/pom.xml
#	pom.xml

+++ catm/src/main/java/com/lt/catm/repositories/UserRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserRepository extends R2dbcRepository<User, Long> {
+public interface UserRepository extends R2dbcRepository<User, Integer> {
 
 }
